Move pipelines/ directory up a level.
