+++ src/invoke_training/pipelines/stable_diffusion/finetune_lora_sd.py
