+++ src/invoke_training/pipelines/stable_diffusion_xl/finetune_lora_and_ti_sdxl.py
@@ -3,42 +3,34 @@ import json
 import logging
 import math
 import os
-import tempfile
 import time
 from pathlib import Path
-from typing import Literal, Optional, Union
+from typing import Literal
 
 import peft
 import torch
 import torch.utils.data
 from accelerate import Accelerator
 from accelerate.utils import set_seed
-from diffusers import AutoencoderKL, DDPMScheduler, UNet2DConditionModel
+from diffusers import UNet2DConditionModel
 from diffusers.optimization import get_scheduler
-from torch.utils.data import DataLoader
 from tqdm.auto import tqdm
-from transformers import CLIPPreTrainedModel, CLIPTextModel, PreTrainedTokenizer
+from transformers import CLIPTextModel
 
-from invoke_training.config.pipelines.finetune_lora_config import FinetuneLoRASDXLConfig
-from invoke_training.config.shared.data.data_loader_config import (
-    DreamboothSDDataLoaderConfig,
-    ImageCaptionSDDataLoaderConfig,
-)
+from invoke_training.config.pipelines.finetune_lora_and_ti_config import FinetuneLoraAndTiSdxlConfig
+from invoke_training.pipelines.stable_diffusion_xl.finetune_lora_sdxl import train_forward
+from invoke_training.pipelines.stable_diffusion_xl.textual_inversion_sdxl import _initialize_placeholder_tokens
 from invoke_training.training._shared.accelerator.accelerator_utils import (
     get_mixed_precision_dtype,
     initialize_accelerator,
     initialize_logging,
 )
 from invoke_training.training._shared.checkpoints.checkpoint_tracker import CheckpointTracker
-from invoke_training.training._shared.data.data_loaders.dreambooth_sd_dataloader import (
-    build_dreambooth_sd_dataloader,
-)
-from invoke_training.training._shared.data.data_loaders.image_caption_sd_dataloader import (
-    build_image_caption_sd_dataloader,
+from invoke_training.training._shared.checkpoints.serialization import save_state_dict
+from invoke_training.training._shared.data.data_loaders.textual_inversion_sd_dataloader import (
+    build_textual_inversion_sd_dataloader,
 )
 from invoke_training.training._shared.data.samplers.aspect_ratio_bucket_batch_sampler import log_aspect_ratio_buckets
-from invoke_training.training._shared.data.transforms.tensor_disk_cache import TensorDiskCache
-from invoke_training.training._shared.data.utils.resolution import Resolution
 from invoke_training.training._shared.optimizer.optimizer_utils import initialize_optimizer
 from invoke_training.training._shared.stable_diffusion.lora_checkpoint_utils import (
     TEXT_ENCODER_TARGET_MODULES,
@@ -46,20 +38,22 @@ from invoke_training.training._shared.stable_diffusion.lora_checkpoint_utils imp
     save_sdxl_kohya_checkpoint,
     save_sdxl_peft_checkpoint,
 )
-from invoke_training.training._shared.stable_diffusion.min_snr_weighting import compute_snr
 from invoke_training.training._shared.stable_diffusion.model_loading_utils import (
     load_models_sdxl,
 )
-from invoke_training.training._shared.stable_diffusion.tokenize_captions import tokenize_captions
+from invoke_training.training._shared.stable_diffusion.textual_inversion import restore_original_embeddings
 from invoke_training.training._shared.stable_diffusion.validation import generate_validation_images_sdxl
-from invoke_training.training.pipelines.stable_diffusion.finetune_lora_sd import cache_vae_outputs
 
 
-def _save_sdxl_lora_checkpoint(
+def _save_sdxl_lora_and_ti_checkpoint(
+    config: FinetuneLoraAndTiSdxlConfig,
     idx: int,
     unet: peft.PeftModel | None,
     text_encoder_1: peft.PeftModel | None,
     text_encoder_2: peft.PeftModel | None,
+    placeholder_token_ids_1: list[int],
+    placeholder_token_ids_2: list[int],
+    accelerator: Accelerator,
     logger: logging.Logger,
     checkpoint_tracker: CheckpointTracker,
     lora_checkpoint_format: Literal["invoke_peft", "kohya"],
@@ -72,252 +66,41 @@ def _save_sdxl_lora_checkpoint(
 
     if lora_checkpoint_format == "invoke_peft":
         save_sdxl_peft_checkpoint(
-            Path(save_path), unet=unet, text_encoder_1=text_encoder_1, text_encoder_2=text_encoder_2
+            Path(save_path),
+            unet=unet if config.train_unet else None,
+            text_encoder_1=text_encoder_1 if config.train_text_encoder else None,
+            text_encoder_2=text_encoder_2 if config.train_text_encoder else None,
         )
     elif lora_checkpoint_format == "kohya":
         save_sdxl_kohya_checkpoint(
-            Path(save_path), unet=unet, text_encoder_1=text_encoder_1, text_encoder_2=text_encoder_2
+            Path(save_path) / "lora.safetensors",
+            unet=unet if config.train_unet else None,
+            text_encoder_1=text_encoder_1 if config.train_text_encoder else None,
+            text_encoder_2=text_encoder_2 if config.train_text_encoder else None,
         )
     else:
         raise ValueError(f"Unsupported lora_checkpoint_format: '{lora_checkpoint_format}'.")
 
-
-def _build_data_loader(
-    data_loader_config: Union[ImageCaptionSDDataLoaderConfig, DreamboothSDDataLoaderConfig],
-    batch_size: int,
-    text_encoder_output_cache_dir: Optional[str] = None,
-    vae_output_cache_dir: Optional[str] = None,
-    shuffle: bool = True,
-    sequential_batching: bool = False,
-) -> DataLoader:
-    if data_loader_config.type == "IMAGE_CAPTION_SD_DATA_LOADER":
-        return build_image_caption_sd_dataloader(
-            config=data_loader_config,
-            batch_size=batch_size,
-            text_encoder_output_cache_dir=text_encoder_output_cache_dir,
-            text_encoder_cache_field_to_output_field={
-                "prompt_embeds": "prompt_embeds",
-                "pooled_prompt_embeds": "pooled_prompt_embeds",
-            },
-            vae_output_cache_dir=vae_output_cache_dir,
-            shuffle=shuffle,
+    if config.train_ti:
+        ti_checkpoint_path = Path(save_path) / "embeddings.safetensors"
+        learned_embeds_1 = (
+            accelerator.unwrap_model(text_encoder_1)
+            .get_input_embeddings()
+            .weight[min(placeholder_token_ids_1) : max(placeholder_token_ids_1) + 1]
         )
-    elif data_loader_config.type == "DREAMBOOTH_SD_DATA_LOADER":
-        return build_dreambooth_sd_dataloader(
-            config=data_loader_config,
-            batch_size=batch_size,
-            text_encoder_output_cache_dir=text_encoder_output_cache_dir,
-            text_encoder_cache_field_to_output_field={
-                "prompt_embeds": "prompt_embeds",
-                "pooled_prompt_embeds": "pooled_prompt_embeds",
-            },
-            vae_output_cache_dir=vae_output_cache_dir,
-            shuffle=shuffle,
-            sequential_batching=sequential_batching,
+        learned_embeds_2 = (
+            accelerator.unwrap_model(text_encoder_2)
+            .get_input_embeddings()
+            .weight[min(placeholder_token_ids_2) : max(placeholder_token_ids_2) + 1]
         )
-    else:
-        raise ValueError(f"Unsupported data loader config type: '{data_loader_config.type}'.")
+        learned_embeds_dict = {
+            "clip_l": learned_embeds_1.detach().cpu(),
+            "clip_g": learned_embeds_2.detach().cpu(),
+        }
+        save_state_dict(learned_embeds_dict, ti_checkpoint_path)
 
 
-# encode_prompt was adapted from:
-# https://github.com/huggingface/diffusers/blob/7b07f9812a58bfa96c06ed8ffe9e6b584286e2fd/examples/text_to_image/train_text_to_image_lora_sdxl.py#L470-L496
-def _encode_prompt(text_encoders: list[CLIPPreTrainedModel], prompt_token_ids_list: list[torch.Tensor]):
-    prompt_embeds_list = []
-
-    for i, text_encoder in enumerate(text_encoders):
-        text_input_ids = prompt_token_ids_list[i]
-
-        prompt_embeds = text_encoder(
-            text_input_ids.to(text_encoder.device),
-            output_hidden_states=True,
-        )
-
-        # We are only ALWAYS interested in the pooled output of the final text encoder.
-        # TODO(ryand): Document this logic more clearly.
-        pooled_prompt_embeds = prompt_embeds[0]
-        prompt_embeds = prompt_embeds.hidden_states[-2]
-        bs_embed, seq_len, _ = prompt_embeds.shape
-        prompt_embeds = prompt_embeds.view(bs_embed, seq_len, -1)
-        prompt_embeds_list.append(prompt_embeds)
-
-    prompt_embeds = torch.concat(prompt_embeds_list, dim=-1)
-    pooled_prompt_embeds = pooled_prompt_embeds.view(bs_embed, -1)
-    return prompt_embeds, pooled_prompt_embeds
-
-
-# TODO(ryand): Cache VAE outputs and text encoder outputs at the same time in a single pass over the dataset.
-
-
-def cache_text_encoder_outputs(
-    cache_dir: str,
-    config: FinetuneLoRASDXLConfig,
-    tokenizer_1: PreTrainedTokenizer,
-    tokenizer_2: PreTrainedTokenizer,
-    text_encoder_1: CLIPPreTrainedModel,
-    text_encoder_2: CLIPPreTrainedModel,
-):
-    """Run the text encoder on all captions in the dataset and cache the results to disk.
-    Args:
-        cache_dir (str): The directory where the results will be cached.
-        config (FinetuneLoRAConfig): Training config.
-        tokenizer_1 (PreTrainedTokenizer):
-        tokenizer_2 (PreTrainedTokenizer):
-        text_encoder_1 (CLIPPreTrainedModel):
-        text_encoder_2 (CLIPPreTrainedModel):
-    """
-    data_loader = _build_data_loader(
-        data_loader_config=config.data_loader,
-        batch_size=config.train_batch_size,
-        shuffle=False,
-        sequential_batching=True,
-    )
-
-    cache = TensorDiskCache(cache_dir)
-
-    for data_batch in tqdm(data_loader):
-        caption_token_ids_1 = tokenize_captions(tokenizer_1, data_batch["caption"])
-        caption_token_ids_2 = tokenize_captions(tokenizer_2, data_batch["caption"])
-        prompt_embeds, pooled_prompt_embeds = _encode_prompt(
-            [text_encoder_1, text_encoder_2], [caption_token_ids_1, caption_token_ids_2]
-        )
-
-        # Split batch before caching.
-        for i in range(len(data_batch["id"])):
-            embeds = {
-                "prompt_embeds": prompt_embeds[i],
-                "pooled_prompt_embeds": pooled_prompt_embeds[i],
-            }
-            cache.save(data_batch["id"][i], embeds)
-
-
-def train_forward(  # noqa: C901
-    accelerator: Accelerator,
-    data_batch: dict,
-    vae: AutoencoderKL,
-    noise_scheduler: DDPMScheduler,
-    tokenizer_1: PreTrainedTokenizer,
-    tokenizer_2: PreTrainedTokenizer,
-    text_encoder_1: CLIPPreTrainedModel,
-    text_encoder_2: CLIPPreTrainedModel,
-    unet: UNet2DConditionModel,
-    weight_dtype: torch.dtype,
-    resolution: int | tuple[int, int],
-    prediction_type=None,
-    min_snr_gamma: float | None = None,
-):
-    """Run the forward training pass for a single data_batch.
-
-    Returns:
-        torch.Tensor: Loss
-    """
-    # Convert images to latent space.
-    # The VAE output may have been cached and included in the data_batch. If not, we calculate it here.
-    latents = data_batch.get("vae_output", None)
-    if latents is None:
-        latents = vae.encode(data_batch["image"].to(dtype=weight_dtype)).latent_dist.sample()
-        latents = latents * vae.config.scaling_factor
-
-    # Sample noise that we'll add to the latents.
-    noise = torch.randn_like(latents)
-
-    batch_size = latents.shape[0]
-    # Sample a random timestep for each image.
-    timesteps = torch.randint(
-        0,
-        noise_scheduler.config.num_train_timesteps,
-        (batch_size,),
-        device=latents.device,
-    )
-    timesteps = timesteps.long()
-
-    # Add noise to the latents according to the noise magnitude at each timestep (this is the forward diffusion
-    # process).
-    noisy_latents = noise_scheduler.add_noise(latents, noise, timesteps)
-
-    # compute_time_ids was copied from:
-    # https://github.com/huggingface/diffusers/blob/7b07f9812a58bfa96c06ed8ffe9e6b584286e2fd/examples/text_to_image/train_text_to_image_lora_sdxl.py#L1033-L1039
-    # "time_ids" may seem like a weird naming choice. The name comes from the diffusers SDXL implementation. Presumably,
-    # it is a result of the fact that the original size and crop values get concatenated with the time embeddings.
-    def compute_time_ids(original_size, crops_coords_top_left):
-        # Adapted from pipeline.StableDiffusionXLPipeline._get_add_time_ids
-        target_size = Resolution.parse(resolution).to_tuple()
-        add_time_ids = list(original_size + crops_coords_top_left + target_size)
-        add_time_ids = torch.tensor([add_time_ids])
-        add_time_ids = add_time_ids.to(accelerator.device, dtype=weight_dtype)
-        return add_time_ids
-
-    add_time_ids = torch.cat(
-        [compute_time_ids(s, c) for s, c in zip(data_batch["original_size_hw"], data_batch["crop_top_left_yx"])]
-    )
-    unet_conditions = {"time_ids": add_time_ids}
-
-    # Get the text embedding for conditioning.
-    # The text encoder output may have been cached and included in the data_batch. If not, we calculate it here.
-    if "prompt_embeds" in data_batch:
-        prompt_embeds = data_batch["prompt_embeds"]
-        pooled_prompt_embeds = data_batch["pooled_prompt_embeds"]
-    else:
-        caption_token_ids_1 = tokenize_captions(tokenizer_1, data_batch["caption"])
-        caption_token_ids_2 = tokenize_captions(tokenizer_2, data_batch["caption"])
-        prompt_embeds, pooled_prompt_embeds = _encode_prompt(
-            [text_encoder_1, text_encoder_2], [caption_token_ids_1, caption_token_ids_2]
-        )
-        prompt_embeds = prompt_embeds.to(dtype=weight_dtype)
-        pooled_prompt_embeds = pooled_prompt_embeds.to(dtype=weight_dtype)
-
-    unet_conditions["text_embeds"] = pooled_prompt_embeds
-
-    # Get the target for loss depending on the prediction type.
-    if prediction_type is not None:
-        # Set the prediction_type of scheduler if it's defined in config.
-        noise_scheduler.register_to_config(prediction_type=prediction_type)
-    if noise_scheduler.config.prediction_type == "epsilon":
-        target = noise
-    elif noise_scheduler.config.prediction_type == "v_prediction":
-        target = noise_scheduler.get_velocity(latents, noise, timesteps)
-    else:
-        raise ValueError(f"Unknown prediction type {noise_scheduler.config.prediction_type}")
-
-    # Predict the noise residual.
-    model_pred = unet(noisy_latents, timesteps, prompt_embeds, added_cond_kwargs=unet_conditions).sample
-
-    min_snr_weights = None
-    if min_snr_gamma is not None:
-        # Compute loss-weights as per Section 3.4 of https://arxiv.org/abs/2303.09556.
-        # Since we predict the noise instead of x_0, the original formulation is slightly changed.
-        # This is discussed in Section 4.2 of the same paper.
-
-        snr = compute_snr(noise_scheduler, timesteps)
-
-        # Note: We divide by snr here per Section 4.2 of the paper, since we are predicting the noise instead of x_0.
-        # w_t = min(1, SNR(t)) / SNR(t)
-        min_snr_weights = torch.clamp(snr, max=min_snr_gamma) / snr
-
-        if noise_scheduler.config.prediction_type == "epsilon":
-            pass
-        elif noise_scheduler.config.prediction_type == "v_prediction":
-            # Velocity objective needs to be floored to an SNR weight of one.
-            min_snr_weights = min_snr_weights + 1
-        else:
-            raise ValueError(f"Unknown prediction type {noise_scheduler.config.prediction_type}")
-
-    loss = torch.nn.functional.mse_loss(model_pred.float(), target.float(), reduction="none")
-
-    # Mean-reduce the loss along all dimensions except for the batch dimension.
-    loss = loss.mean(dim=list(range(1, len(loss.shape))))
-
-    # Apply min_snr_weights.
-    if min_snr_weights is not None:
-        loss = loss * min_snr_weights
-
-    # Apply per-example loss weights.
-    if "loss_weight" in data_batch:
-        loss = loss * data_batch["loss_weight"]
-
-    return loss.mean()
-
-
-def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
+def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
     # Give a clear error message if an unsupported base model was chosen.
     # TODO(ryan): Update this check to work with single-file SD checkpoints.
     # check_base_model_version(
@@ -358,7 +141,6 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
         model_name_or_path=config.model,
         hf_variant=config.hf_variant,
         vae_model=config.vae_model,
-        base_embeddings=config.base_embeddings,
     )
 
     if config.xformers:
@@ -370,31 +152,9 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
         vae.enable_xformers_memory_efficient_attention()
 
     # Prepare text encoder output cache.
-    text_encoder_output_cache_dir_name = None
+    # text_encoder_output_cache_dir_name = None
     if config.cache_text_encoder_outputs:
-        # TODO(ryand): Think about how to better check if it is safe to cache the text encoder outputs. Currently, there
-        # are a number of configurations that would cause variation in the text encoder outputs and should not be used
-        # with caching.
-
-        if config.train_text_encoder:
-            raise ValueError("'cache_text_encoder_outputs' and 'train_text_encoder' cannot both be True.")
-
-        # We use a temporary directory for the cache. The directory will automatically be cleaned up when
-        # tmp_text_encoder_output_cache_dir is destroyed.
-        tmp_text_encoder_output_cache_dir = tempfile.TemporaryDirectory()
-        text_encoder_output_cache_dir_name = tmp_text_encoder_output_cache_dir.name
-        if accelerator.is_local_main_process:
-            # Only the main process should populate the cache.
-            logger.info(f"Generating text encoder output cache ('{text_encoder_output_cache_dir_name}').")
-            text_encoder_1.to(accelerator.device, dtype=weight_dtype)
-            text_encoder_2.to(accelerator.device, dtype=weight_dtype)
-            cache_text_encoder_outputs(
-                text_encoder_output_cache_dir_name, config, tokenizer_1, tokenizer_2, text_encoder_1, text_encoder_2
-            )
-        # Move the text_encoders back to the CPU, because they are not needed for training.
-        text_encoder_1.to("cpu")
-        text_encoder_2.to("cpu")
-        accelerator.wait_for_everyone()
+        raise NotImplementedError("Caching text encoder outputs is not yet supported.")
     else:
         text_encoder_1.to(accelerator.device, dtype=weight_dtype)
         text_encoder_2.to(accelerator.device, dtype=weight_dtype)
@@ -402,29 +162,7 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
     # Prepare VAE output cache.
     vae_output_cache_dir_name = None
     if config.cache_vae_outputs:
-        if config.data_loader.random_flip:
-            raise ValueError("'cache_vae_outputs' cannot be True if 'random_flip' is True.")
-        if not config.data_loader.center_crop:
-            raise ValueError("'cache_vae_outputs' cannot be True if 'center_crop' is False.")
-
-        # We use a temporary directory for the cache. The directory will automatically be cleaned up when
-        # tmp_vae_output_cache_dir is destroyed.
-        tmp_vae_output_cache_dir = tempfile.TemporaryDirectory()
-        vae_output_cache_dir_name = tmp_vae_output_cache_dir.name
-        if accelerator.is_local_main_process:
-            # Only the main process should to populate the cache.
-            logger.info(f"Generating VAE output cache ('{vae_output_cache_dir_name}').")
-            vae.to(accelerator.device, dtype=weight_dtype)
-            data_loader = _build_data_loader(
-                data_loader_config=config.data_loader,
-                batch_size=config.train_batch_size,
-                shuffle=False,
-                sequential_batching=True,
-            )
-            cache_vae_outputs(vae_output_cache_dir_name, data_loader, vae)
-        # Move the VAE back to the CPU, because it is not needed for training.
-        vae.to("cpu")
-        accelerator.wait_for_everyone()
+        raise NotImplementedError("Caching VAE outputs is not yet supported.")
     else:
         vae.to(accelerator.device, dtype=weight_dtype)
 
@@ -432,23 +170,19 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
 
     # Add LoRA layers to the models being trained.
     trainable_param_groups = []
-    all_trainable_models: list[peft.PeftModel] = []
+    all_trainable_models: set[torch.nn.Module] = set()
 
-    def inject_lora_layers(model, lora_config: peft.LoraConfig, lr: float | None = None) -> peft.PeftModel:
+    def inject_lora_layers(model, lora_config: peft.LoraConfig, lr: float) -> peft.PeftModel:
         peft_model = peft.get_peft_model(model, lora_config)
         peft_model.print_trainable_parameters()
 
         # Populate `trainable_param_groups`, to be passed to the optimizer.
-        param_group = {"params": list(filter(lambda p: p.requires_grad, peft_model.parameters()))}
-        if lr is not None:
-            param_group["lr"] = lr
+        param_group = {"params": list(filter(lambda p: p.requires_grad, peft_model.parameters())), "lr": lr}
         trainable_param_groups.append(param_group)
 
         # Populate all_trainable_models.
-        all_trainable_models.append(peft_model)
-
+        all_trainable_models.add(peft_model)
         peft_model.train()
-
         return peft_model
 
     if config.train_unet:
@@ -473,6 +207,31 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
         text_encoder_2 = inject_lora_layers(
             text_encoder_2, text_encoder_lora_config, lr=config.text_encoder_learning_rate
         )
+
+    if config.train_ti:
+        # TODO(ryand): Move this private function to a shared location.
+        placeholder_tokens, placeholder_token_ids_1, placeholder_token_ids_2 = _initialize_placeholder_tokens(
+            config=config,
+            tokenizer_1=tokenizer_1,
+            tokenizer_2=tokenizer_2,
+            text_encoder_1=text_encoder_1,
+            text_encoder_2=text_encoder_2,
+        )
+        logger.info(f"Initialized {len(placeholder_tokens)} placeholder tokens: {placeholder_tokens}.")
+
+        # Unfreeze the token embeddings in the text encoders.
+        text_encoder_1.text_model.embeddings.token_embedding.requires_grad_(True)
+        text_encoder_2.text_model.embeddings.token_embedding.requires_grad_(True)
+
+        all_trainable_models.add(text_encoder_1)
+        all_trainable_models.add(text_encoder_2)
+
+        for te in [text_encoder_1, text_encoder_2]:
+            param_group = {
+                "params": te.get_input_embeddings().parameters(),
+                "lr": config.textual_inversion_learning_rate,
+            }
+            trainable_param_groups.append(param_group)
 
     # Make sure all trainable params are in float32.
     for trainable_model in all_trainable_models:
@@ -500,15 +259,15 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
 
                 # Set requires_grad = True on the first parameters of the text encoders. Without this, the text encoder
                 # LoRA weights would have 0 gradients, and so would not get trained. Note that the set of
-                # trainable_param_groups has already been populated - the embeddings will not be trained.
+                # trainable_param_groups has already been populated - this won't change what gets trained.
                 te.text_model.embeddings.requires_grad_(True)
 
     optimizer = initialize_optimizer(config.optimizer, trainable_param_groups)
 
-    data_loader = _build_data_loader(
-        data_loader_config=config.data_loader,
+    data_loader = build_textual_inversion_sd_dataloader(
+        config=config.data_loader,
+        placeholder_token=config.placeholder_token,
         batch_size=config.train_batch_size,
-        text_encoder_output_cache_dir=text_encoder_output_cache_dir_name,
         vae_output_cache_dir=vae_output_cache_dir_name,
     )
 
@@ -564,7 +323,7 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
     unet, text_encoder_1, text_encoder_2, optimizer, data_loader, lr_scheduler = prepared_result
 
     if accelerator.is_main_process:
-        accelerator.init_trackers("lora_training")
+        accelerator.init_trackers("lora_and_ti_training")
         # Tensorboard uses markdown formatting, so we wrap the config json in a code block.
         accelerator.log({"configuration": f"```json\n{json.dumps(config.dict(), indent=2, default=str)}\n```\n"})
 
@@ -572,14 +331,12 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
         base_dir=ckpt_dir,
         prefix="checkpoint_epoch",
         max_checkpoints=config.max_checkpoints,
-        extension=".safetensors" if config.lora_checkpoint_format == "kohya" else None,
     )
 
     step_checkpoint_tracker = CheckpointTracker(
         base_dir=ckpt_dir,
         prefix="checkpoint_step",
         max_checkpoints=config.max_checkpoints,
-        extension=".safetensors" if config.lora_checkpoint_format == "kohya" else None,
     )
 
     # Train!
@@ -603,9 +360,34 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
     )
     progress_bar.set_description("Steps")
 
+    ti_train_steps = num_train_steps
+    if config.ti_train_steps_ratio is not None:
+        ti_train_steps = math.ceil(num_train_steps * config.ti_train_steps_ratio)
+        logger.info(f"The TI training pivot point is set at {ti_train_steps} steps.")
+
+    # Keep original embeddings as reference.
+    with torch.no_grad():
+        orig_embeds_params_1 = accelerator.unwrap_model(text_encoder_1).get_input_embeddings().weight.data.clone()
+        orig_embeds_params_2 = accelerator.unwrap_model(text_encoder_2).get_input_embeddings().weight.data.clone()
+
     for epoch in range(first_epoch, num_train_epochs):
+        # TODO(ryand): Is this necessary?
+        text_encoder_1.train()
+        text_encoder_2.train()
+
         train_loss = 0.0
         for data_batch in data_loader:
+            if global_step == ti_train_steps and config.train_ti:
+                logger.info("Reached TI training pivot point. Setting TI learning rate to 0.0.")
+                # TODO(ryand): The TI embeddings continue to be updated slightly by the normalization step in
+                # restore_original_embeddings(...). The updates should be very small and converge quickly, so this
+                # should be fine. But, at some point we should tidy this up.
+                for ti_param_group in optimizer.param_groups[-2:]:
+                    # The TI param groups should be the last two param groups. But, this is pretty brittle, so this
+                    # assertion adds a bit of safety.
+                    assert len(ti_param_group["params"]) == 1
+                    ti_param_group["lr"] = 0.0
+
             with accelerator.accumulate(unet, text_encoder_1, text_encoder_2):
                 loss = train_forward(
                     accelerator=accelerator,
@@ -637,6 +419,23 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
                 lr_scheduler.step()
                 optimizer.zero_grad(set_to_none=True)
 
+                # Make sure we don't update any embedding weights besides the newly-added token(s).
+                # TODO(ryand): Should we only do this if accelerator.sync_gradients?
+                restore_original_embeddings(
+                    tokenizer=tokenizer_1,
+                    placeholder_token_ids=placeholder_token_ids_1,
+                    accelerator=accelerator,
+                    text_encoder=text_encoder_1,
+                    orig_embeds_params=orig_embeds_params_1,
+                )
+                restore_original_embeddings(
+                    tokenizer=tokenizer_2,
+                    placeholder_token_ids=placeholder_token_ids_2,
+                    accelerator=accelerator,
+                    text_encoder=text_encoder_2,
+                    orig_embeds_params=orig_embeds_params_2,
+                )
+
             # Checks if the accelerator has performed an optimization step behind the scenes.
             if accelerator.sync_gradients:
                 progress_bar.update(1)
@@ -644,16 +443,25 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
                 log = {"train_loss": train_loss}
 
                 lrs = lr_scheduler.get_last_lr()
+
+                # Prepare LR names in the same order that their respective param groups were added to the optimizer.
+                # TODO: Do this at the time that we prepare the param groups?
+                lr_names = []
                 if config.train_unet:
-                    # When training the UNet, it will always be the first parameter group.
-                    log["lr/unet"] = float(lrs[0])
-                    if config.optimizer.optimizer_type == "Prodigy":
-                        log["lr/d*lr/unet"] = optimizer.param_groups[0]["d"] * optimizer.param_groups[0]["lr"]
+                    lr_names.append("unet")
                 if config.train_text_encoder:
-                    # When training the text encoder, it will always be the last parameter group.
-                    log["lr/text_encoder"] = float(lrs[-1])
+                    lr_names.append("text_encoder_1")
+                    lr_names.append("text_encoder_2")
+                if config.train_ti:
+                    lr_names.append("ti_embeddings_1")
+                    lr_names.append("ti_embeddings_2")
+
+                for lr_idx, lr_name in enumerate(lr_names):
+                    log[f"lr/{lr_name}"] = float(lrs[lr_idx])
                     if config.optimizer.optimizer_type == "Prodigy":
-                        log["lr/d*lr/text_encoder"] = optimizer.param_groups[-1]["d"] * optimizer.param_groups[-1]["lr"]
+                        log[f"lr/d*lr/{lr_name}"] = (
+                            optimizer.param_groups[lr_idx]["d"] * optimizer.param_groups[lr_idx]["lr"]
+                        )
 
                 accelerator.log(log, step=global_step)
                 train_loss = 0.0
@@ -661,11 +469,15 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
                 if config.save_every_n_steps is not None and (global_step + 1) % config.save_every_n_steps == 0:
                     accelerator.wait_for_everyone()
                     if accelerator.is_main_process:
-                        _save_sdxl_lora_checkpoint(
+                        _save_sdxl_lora_and_ti_checkpoint(
+                            config=config,
                             idx=global_step + 1,
                             unet=unet,
                             text_encoder_1=text_encoder_1,
                             text_encoder_2=text_encoder_2,
+                            placeholder_token_ids_1=placeholder_token_ids_1,
+                            placeholder_token_ids_2=placeholder_token_ids_2,
+                            accelerator=accelerator,
                             logger=logger,
                             checkpoint_tracker=step_checkpoint_tracker,
                             lora_checkpoint_format=config.lora_checkpoint_format,
@@ -693,7 +505,6 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
                             logger=logger,
                             prefix="step",
                         )
-
             logs = {
                 "step_loss": loss.detach().item(),
                 "lr": lr_scheduler.get_last_lr()[0],
@@ -706,11 +517,15 @@ def run_training(config: FinetuneLoRASDXLConfig):  # noqa: C901
         # Save a checkpoint every n epochs.
         if config.save_every_n_epochs is not None and (epoch + 1) % config.save_every_n_epochs == 0:
             if accelerator.is_main_process:
-                _save_sdxl_lora_checkpoint(
+                _save_sdxl_lora_and_ti_checkpoint(
+                    config=config,
                     idx=epoch + 1,
                     unet=unet,
                     text_encoder_1=text_encoder_1,
                     text_encoder_2=text_encoder_2,
+                    placeholder_token_ids_1=placeholder_token_ids_1,
+                    placeholder_token_ids_2=placeholder_token_ids_2,
+                    accelerator=accelerator,
                     logger=logger,
                     checkpoint_tracker=epoch_checkpoint_tracker,
                     lora_checkpoint_format=config.lora_checkpoint_format,
