+++ src/invoke_training/scripts/invoke_train.py
@@ -5,21 +5,21 @@ import yaml
 from pydantic import TypeAdapter
 
 from invoke_training.config.pipelines.pipeline_config import PipelineConfig
-from invoke_training.training._experimental.dpo.diffusion_dpo_lora_sd import (
-    run_training as run_diffusion_dpo_sd,
-)
-from invoke_training.training.pipelines.stable_diffusion.finetune_lora_sd import run_training as run_finetune_lora_sd
-from invoke_training.training.pipelines.stable_diffusion.textual_inversion_sd import (
+from invoke_training.pipelines.stable_diffusion.finetune_lora_sd import run_training as run_finetune_lora_sd
+from invoke_training.pipelines.stable_diffusion.textual_inversion_sd import (
     run_training as run_textual_inversion_sd,
 )
-from invoke_training.training.pipelines.stable_diffusion_xl.finetune_lora_and_ti_sdxl import (
+from invoke_training.pipelines.stable_diffusion_xl.finetune_lora_and_ti_sdxl import (
     run_training as run_finetune_lora_and_ti_sdxl,
 )
-from invoke_training.training.pipelines.stable_diffusion_xl.finetune_lora_sdxl import (
+from invoke_training.pipelines.stable_diffusion_xl.finetune_lora_sdxl import (
     run_training as run_finetune_lora_sdxl,
 )
-from invoke_training.training.pipelines.stable_diffusion_xl.textual_inversion_sdxl import (
+from invoke_training.pipelines.stable_diffusion_xl.textual_inversion_sdxl import (
     run_training as run_textual_inversion_sdxl,
+)
+from invoke_training.training._experimental.dpo.diffusion_dpo_lora_sd import (
+    run_training as run_diffusion_dpo_sd,
 )
